fix: recipes API handling
